Fix signed/unsigned comparison in assertion

+++ groups/bdl/bdldfp/bdldfp_decimalconvertutil.cpp
@@ -233,8 +233,8 @@ void restoreDecimalFromBinary(DECIMAL_TYPE *dfp, BINARY_TYPE bfp)
     char buffer[48];
     int rc = snprintf(buffer, sizeof(buffer), StdioFormat<BINARY_TYPE>::format(), bfp);
     (void)rc;
-    BSLS_ASSERT(0 <= rc && rc < sizeof(buffer));   
-    
+    BSLS_ASSERT(0 <= rc && static_cast<unsigned>(rc) < sizeof(buffer));
+
     typename DecimalTraits<DECIMAL_TYPE>::SignificandType significand(0);
     int  exponent(0);
     bool negative(false);
